requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==2.2.5
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.9.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.1
ml-dtypes==0.2.0
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
Pillow==9.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.14.1
gunicorn